refactor: move pg concerns into edm
